use new test helper in other tests
